fix: errors in the new pid issuance flow
